pyproject.toml
[project]
name = "vrchat-announce-bot"
version = "0.1.0"
description = "Discord bot to schedule VRChat group announcements"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]

dependencies = [
    "discord.py>=2.3.0",
    "openai>=1.0.0",
    "PyYAML>=6.0",
    "APScheduler>=3.10.0",
    "vrchatapi>=1.19.1",
    "python-dotenv>=1.0.0",
    "pytz>=2025.2",
    "python-dateutil>=2.9.0.post0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
